estate_project/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
fonttools==4.39.4
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.1
pyparsing==3.0.9
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
